requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.1
SQLAlchemy==2.0.47
typing_extensions==4.15.0
Werkzeug==3.0.3